feat: show warning about private profile when there are 0 stats
Future improvement:
Right now, in their api, github does not expose the information
about whether the user has made there profile private or not. If
they do so, we can update the logic to avoid making stats API call
if the profile is private and the user's token is not there.